moved menu into header
