+++ src/acf/group_645b71459f7b3.json
@@ -3,8 +3,8 @@
     "title": "Footer",
     "fields": [
         {
-            "key": "field_645b87a15dc95",
-            "label": "Footer",
+            "key": "field_645ce496b04f8",
+            "label": "Cloned",
             "name": "footer",
             "aria-label": "",
             "type": "clone",
@@ -17,53 +17,14 @@
                 "id": ""
             },
             "clone": [
-                "field_645b878634d16",
+                "field_645ce5d7fc9da",
                 "field_645b8c5edb7a9",
-                "field_5b868d1e9ea0d"
-            ],
-            "display": "seamless",
-            "layout": "block",
-            "prefix_label": 0,
-            "prefix_name": 1
-        },
-        {
-            "key": "field_645b8cd82ca11",
-            "label": "Copyright",
-            "name": "footer",
-            "aria-label": "",
-            "type": "clone",
-            "instructions": "",
-            "required": 0,
-            "conditional_logic": 0,
-            "wrapper": {
-                "width": "",
-                "class": "",
-                "id": ""
-            },
-            "clone": [
-                "field_64368ce753370"
-            ],
-            "display": "seamless",
-            "layout": "block",
-            "prefix_label": 1,
-            "prefix_name": 1
-        },
-        {
-            "key": "field_645b87e85dc96",
-            "label": "Styles",
-            "name": "footer",
-            "aria-label": "",
-            "type": "clone",
-            "instructions": "",
-            "required": 0,
-            "conditional_logic": 0,
-            "wrapper": {
-                "width": "",
-                "class": "",
-                "id": ""
-            },
-            "clone": [
+                "field_645ce607fc9dc",
+                "field_5b868d1e9ea0d",
+                "field_645ce64e226b3",
+                "field_64368ce753370",
                 "field_64242b58f9e08",
+                "field_64242c4c0aed4",
                 "field_641d84ec7d6f1",
                 "field_641d8315cc2fe",
                 "field_644655d072e09"
@@ -92,5 +53,5 @@
     "active": true,
     "description": "",
     "show_in_rest": 0,
-    "modified": 1683724111
+    "modified": 1683810289
 }
