+++ src/js/components/header/header.module.scss.d.ts
@@ -1,11 +1,31 @@
 // This file is automatically generated.
 // Please do not change this file!
 interface CssExports {
+  'angle-close': string;
+  'angle-open': string;
+  'animation-color-1': string;
+  'animation-color-2': string;
+  'animation-color-3': string;
+  'animation-color-4': string;
+  'animation-dark': string;
+  'animation-light': string;
   'cta': string;
   'cta__link': string;
   'header': string;
   'header__logo': string;
+  'is-active': string;
   'is-hidden': string;
+  'kf-angle-close': string;
+  'kf-angle-open': string;
+  'menu': string;
+  'nav': string;
+  'nav-open': string;
+  'sub-nav': string;
+  'sub-nav-visible': string;
+  'trigger': string;
+  'trigger__bar': string;
+  'trigger__container': string;
+  'trigger__title': string;
 }
 export const cssExports: CssExports;
 export default cssExports;
