--- src/js/components/menu/menu.module.scss.d.ts
@@ -1,18 +1,0 @@
-// This file is automatically generated.
-// Please do not change this file!
-interface CssExports {
-  'animation-color-1': string;
-  'animation-color-2': string;
-  'animation-color-3': string;
-  'animation-color-4': string;
-  'animation-dark': string;
-  'animation-light': string;
-  'menu': string;
-  'nav': string;
-  'nav-open': string;
-  'sub-nav': string;
-  'sub-nav-visible': string;
-  'trigger': string;
-}
-export const cssExports: CssExports;
-export default cssExports;
